// frontend/src/legacy/api/types.ts
export interface AdminVrfConfig {
  maxGasPrice: string;
  maxGasLimit: string;
  callbackGasPrice: string;
  callbackGasLimit: string;
  requestedRngCount: number;
  clientSeed: number;
  lastConfiguredAt: string;
}